manager/pyproject.toml
[project]
name = "manager"
version = "0.0.0"
requires-python = ">=3.13"
dependencies = ["pydantic~=2.10.2", "GitPython~=3.1.43", "PyGithub~=2.5.0"]

[project.optional-dependencies]
dev = ["pytest~=8.3.3", "ruff~=0.8.1"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
